+++ service-tree/src/test/java/com/zhuinden/servicetree/TraversalTest.java
@@ -56,17 +56,17 @@ public class TraversalTest {
         serviceTree = new ServiceTree();
 
         root1Node = serviceTree.createRootNode(root1);
-        child1ANode = serviceTree.createChildNode(root1Node, child1A);
-        child1BNode = serviceTree.createChildNode(root1Node, child1B);
+        child1ANode = root1Node.createChild(child1A);
+        child1BNode = root1Node.createChild(child1B);
 
         root2Node = serviceTree.createRootNode(root2);
-        child2ANode = serviceTree.createChildNode(root2Node, child2A);
-        child2BNode = serviceTree.createChildNode(root2Node, child2B);
+        child2ANode = root2Node.createChild(child2A);
+        child2BNode = root2Node.createChild(child2B);
 
-        child1A1Node = serviceTree.createChildNode(child1ANode, child1A1);
-        child1A2Node = serviceTree.createChildNode(child1ANode, child1A2);
-        child1B1Node = serviceTree.createChildNode(child1BNode, child1B1);
-        child1B2Node = serviceTree.createChildNode(child1BNode, child1B2);
+        child1A1Node = child1ANode.createChild(child1A1);
+        child1A2Node = child1ANode.createChild(child1A2);
+        child1B1Node = child1BNode.createChild(child1B1);
+        child1B2Node = child1BNode.createChild(child1B2);
     }
 
     @Test